Fix (hopefully) piety display bug

+++ src/xtra1.c
@@ -146,7 +146,8 @@ static void prt_piety(void)
 
 	c_put_str(TERM_L_WHITE, "Pt ", ROW_PIETY, COL_PIETY);
 
-	sprintf(tmp, "%9ld", p_ptr->grace);
+	/* apparently this can bug out if it's not a signed long int --Amy */
+	sprintf(tmp, "%9ld", (signed long int)p_ptr->grace);
 
 	c_put_str((p_ptr->praying) ? TERM_L_BLUE : TERM_GREEN, tmp, ROW_PIETY,
 		COL_PIETY + 3);
